Implemented obstacles & made some alterations to code. Too big of a c…

+++ src/Level/Level.cpp
@@ -1,6 +1,7 @@
 #include "Level.h"
 #include <utility>
 #include <algorithm>
+
 Level::Level(std::vector<TileLayer>&& tileLayers, std::vector<sf::Vector2i>&& collisionLayer,
 	const std::string& levelName, const sf::Vector2i& levelSize, int tileSize)
 	: m_tileLayers(std::move(tileLayers)),
@@ -19,11 +20,11 @@ void Level::draw(sf::RenderWindow& window)
 	}
 }
 
-bool Level::collisionTileAtPosition(const sf::Vector2f & position) const
+bool Level::collisionTileAtPosition(const sf::Vector2i & position) const
 {
 	for (const auto& collidableTile : m_collisionLayer)
 	{
-		const sf::Vector2f tilePosition(std::floor(collidableTile.x / m_tileSize), std::floor(collidableTile.y / m_tileSize));
+		const sf::Vector2i tilePosition(std::floor(collidableTile.x / m_tileSize), std::floor(collidableTile.y / m_tileSize));
 		if (position == tilePosition)
 		{
 			return true;
